Fix tasks.json definition for Cursor
We were relying on the extension connor4312.esbuild-problem-matchers, but its not in the OpenVSX marketplace, so cursor couldnt install it. This meant that when building/developing the extension with cursor it wouldnt startup automatically because it didnt detect the `[watch] build finished` output. Now it does

also updated some package.json commands to use yarn as that is our package manager.